move tests to external folder.
