--- client/examples/ask.rs
@@ -1,4 +1,0 @@
-// Copyright 2020 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-fn main() {}
